pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Agentic deep research assistant with iterative search, reflection, and report generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "openai>=1.0",
    "tenacity>=8.2",
    "tavily-python>=0.3",
    "tiktoken>=0.6",
    "python-dotenv>=1.0",
    "typing-extensions>=4.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["agent*", "backend*"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
markers = ["integration: tests that hit real LLM / network services"]
addopts = "-m 'not integration' -q"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
